# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JABEEN Investor Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (managers+)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [investor, top-management, project-manager, administrator]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user account (PM creates investors; Admin creates any)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user with one-time password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreated"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user detail (managers+)
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user account (admin only)
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user account (admin only)
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "204":
          description: Deleted
  /users/{userId}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password (admin only)
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: New one-time password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordReset"

  # --- AUDIT LOG ---
...
components:
  parameters:
...
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    User:
      type: object
      required: [id, email, fullName, companyName, role, active, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        companyName:
          type: string
        title:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [investor, top-management, project-manager, administrator]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [fullName, email, companyName, role]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 160
        email:
          type: string
          format: email
        companyName:
          type: string
          minLength: 2
          maxLength: 200
        title:
          type: string
          maxLength: 160
        phone:
          type: string
          maxLength: 40
        role:
          type: string
          enum: [investor, top-management, project-manager, administrator]
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 160
        companyName:
          type: string
          minLength: 2
          maxLength: 200
        title:
          type: ["string", "null"]
          maxLength: 160
        phone:
          type: ["string", "null"]
          maxLength: 40
        role:
          type: string
          enum: [investor, top-management, project-manager, administrator]
        active:
          type: boolean
    UserCreated:
      type: object
      required: [user, temporaryPassword]
      properties:
        user:
          $ref: "#/components/schemas/User"
        temporaryPassword:
          type: string
    PasswordReset:
      type: object
      required: [temporaryPassword]
      properties:
        temporaryPassword:
          type: string

    # --- PROJECT ---
...